backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic[v2]>=2.7.0
langgraph>=0.1.0
sqlmodel>=0.0.18
httpx>=0.27.0
tenacity>=8.3.0
structlog>=24.1.0
instructor>=1.3.0
docling>=1.0.0
pymupdf>=1.24.0
pillow>=10.3.0
opencv-python>=4.9.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
opentelemetry-sdk>=1.24.0
arize-phoenix>=4.0.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
groq>=0.9.0
PyJWT==2.8.0
cryptography==42.0.5
